Added downsampling patch operator.
Also fixed INTER_AREA to INTER_NEAREST and removed remaining TODOs, since that is the correct interpolation method to be using.
TODO: currently only works on 3x3 patches and scale = 2. This needs to be modified.
TODO: test fails for new ApplyToPixel method. Downsampling operator needs to be fully tested in this case.