Made TV Regularizer faster.
Replaced the O(n^2) code that checked every single possible partial derivative with code that only checks the partials that can possibly have values.
TODO: verify that FADBAD++ is actually doing this in O(n) time, since if the representation is sparse there may be more complexity in doing it this way if the library is implemented badly.