Minor naming/print fixes, new test motion file.
Fixed partial derivative variable names in TV Regularizer.
The SuperResolution binary now prints the number of observations it has to work with.
Added the --number_of_frames parameter to run.py when running SuperResolution.
Added a new motion sequence file, this one with 9 different shifts for up to 3X super-resolution.