ImageData now respects luminance_channel_only_.
The interpolation to convert back to BGR is still not working. Also need to add tests.
TODO: when converting back to BGR, the non-luminance channels need to be interpolated to the correct size (if the image was resized).
TODO: allow fusing a monochrome modified image and interpolate in the color information from another image.