Added preprocessor macro for root code directory.
This can be used to convert relative paths within the code directory to an absolute path for accessing test data.
All unit tests that use relative paths to access data have been updated accordingly.
CMake will define this flag, but if for some reason it is not available, the GetRootCodeDirectory() function will still compile but only return the local Unix directory ("."). This will, however, cause unit tests to fail.