Updated DisplayImage utility.
It now has different dimensions for width and height, and adapts scaling to always fit the maximum available sizes in either dimension.
It can also dynamically decide whether to scale the image up or down, based on its current size (before it only scaled up).
Also updated the VisualizeImage binary to support rescaling as a parameter.