Interpolated image is not created unless needed.
The linearly-upsampled image will only be created if needed for visualization (comparison) or evaluation. Otherwise, it wastes memory and is never used.
Also added a check for constraining the max number of SR solver parameters that can be super-resolved to the max value of an int.
TODO: regularizers create new vectors for the values and gradients, which takes up memory. This should be fixed by passing in the solver's array instead and directly modifying that.