UNSTABLE COMMIT: IRLS solver can split channels.
Some parts of the code are commented out because other classes need to be modified to support this changes. These will be modified in the next few commits.
Right now the code compiles, but it will fail to solve/pass tests until the other classes are fixed and the modified pieces are uncommented.