UNSTABLE COMMIT: AddChannel method using double*.
AddChannel(const double*, const cv::Size&) now in ImageData.
Before the only way to add a channel was using a cv::Mat. Now it can be added using a data array and a cv::Size.
Uncommented next part of the IRLS solver that required this method, but it still does not do any solving. Need to fix other classes next.