Regularizer::ApplyToImage* now takes the number of channels as argument.
Before this was a class field, but that required updating the class before solving. This was sloppy, so now it's just passed in as an argument.
Also updated tests and the way visualization is handled in test_map_solver.
TODO: Test fails on split vs. non-split TV regularization. This needs to be debugged.